Fix tab close and widget delete
Seems the tab index updates when tabs are removed or inserted
around them. Thus, using the int we get back from the
->addTab call as an index to a map.

Instead, we can use the ->widget(tabIndex) call on the Tab widget,
and then, _after we've removed the tab_, we can then delete the
widget that is the TextBrowser widget.

Note, this MIGHT cause issue if the tab is closed (and the the
text browser, thus deleted) before a network request comes
back. As the tab browser widget is used as the "originating
object" for the requests.